# BackEnd/tests/_regression_snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "BoughtBrrrDealRes": {
        "properties": {
          "HMLInterestRate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Interest paid during HML period (cash)",
            "title": "Hmlinterestrate"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Property address",
            "title": "Address"
          },
          "annual_insurance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "description": "Annual insurance expense",
            "title": "Annual Insurance"
          },
          "annual_property_taxes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "description": "Annual property taxes",
            "title": "Annual Property Taxes"
          },
          "arv_in_thousands": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0",
            "description": "ARV in thousands",
            "title": "Arv In Thousands"
          },
          "bathrooms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Number of bathrooms",
            "title": "Bathrooms"
          },
          "bedrooms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Number of bedrooms",
            "title": "Bedrooms"
          },
          "boughtStage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "purchase",
            "title": "Boughtstage"
          },
          "breakdowns": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "$ref": "#/components/schemas/CalcStep"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breakdowns"
          },
          "capexPercent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "title": "Capexpercent"
          },
          "cashReserve": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "title": "Cashreserve"
          },
          "cash_flow": {
            "title": "Cash Flow",
            "type": "number"
          },
          "cash_on_cash": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash On Cash"
          },
          "cash_out": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Out"
          },
          "cash_out_routi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Out Routi"
          },
          "closingCostsBuy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "description": "Closing costs when purchasing",
            "title": "Closingcostsbuy"
          },
          "closingCostsRefi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "title": "Closingcostsrefi"
          },
          "completedSubstages": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "boolean"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": {},
            "title": "Completedsubstages"
          },
          "contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "crime_rate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Rate"
          },
          "daysUntilRefi": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 180,
            "title": "Daysuntilrefi"
          },
          "deal_type": {
            "const": "BRRRR",
            "default": "BRRRR",
            "title": "Deal Type",
            "type": "string"
          },
          "down_payment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Down payment percentage (0-100)",
            "title": "Down Payment"
          },
          "dscr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dscr"
          },
          "equity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equity"
          },
          "hmlPoints": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "description": "Hard money lender points (percentage)",
            "title": "Hmlpoints"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interestRate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interestrate"
          },
          "loanTermYears": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 30,
            "title": "Loantermyears"
          },
          "ltv_as_precent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "LTV for Refi",
            "title": "Ltv As Precent"
          },
          "maintenancePercent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "title": "Maintenancepercent"
          },
          "messages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messages"
          },
          "montly_hoa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "description": "Monthly HOA dues",
            "title": "Montly Hoa"
          },
          "net_profit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Profit"
          },
          "niche": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Niche"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "overall_design": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Design"
          },
          "pics_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pics Link"
          },
          "property_managment_fee_precentages_from_rent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "title": "Property Managment Fee Precentages From Rent"
          },
          "purchasePrice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Acquisition price for the property",
            "title": "Purchaseprice"
          },
          "refiPoints": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "2",
            "title": "Refipoints"
          },
          "rehabContingency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "description": "Contingency percentage for rehab costs",
            "title": "Rehabcontingency"
          },
          "rehabCost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "description": "Estimated rehab costs included in the deal",
            "title": "Rehabcost"
          },
          "rent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Expected monthly rent",
            "title": "Rent"
          },
          "rent_comps": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RentComp-Output"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rent Comps"
          },
          "roi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roi"
          },
          "section": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Section number",
            "title": "Section"
          },
          "sold_comps": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SoldComp-Output"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sold Comps"
          },
          "sourceDealId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcedealid"
          },
          "sqft": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Property square footage",
            "title": "Sqft"
          },
          "stage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Stage number",
            "title": "Stage"
          },
          "task": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task"
          },
          "total_cash_needed_for_deal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cash Needed For Deal"
          },
          "total_cash_needed_for_deal_with_buffer": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cash Needed For Deal With Buffer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "use_HM_for_rehab": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "description": "",
            "title": "Use Hm For Rehab"
          },
          "vacancyPercent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.0",
            "title": "Vacancypercent"
          },
          "zillow_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zillow Link"
          }
        },
        "required": [
          "cash_flow",
          "section",
          "stage",
          "address",
          "ltv_as_precent",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "BoughtBrrrDealRes",
        "type": "object"
      },
...
      "CalcStep": {
        "description": "A single self-documenting line in a calculation.\n\n`label`   - human-readable name of the variable (\"Operating Expenses\").\n`value`   - the numeric result of the step (always a float for JSON).\n`formula` - the literal expression evaluated, with concrete values\n            substituted (e.g. \"$2000 - $800 - $1000 = $200\").",
        "properties": {
          "formula": {
            "title": "Formula",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "label",
          "value",
          "formula"
        ],
        "title": "CalcStep",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PipelineTemplateRes": {
        "description": "Response shape for a single pipeline template row.",
        "properties": {
          "dealType": {
            "enum": [
              "BRRRR",
              "FLIP"
            ],
            "title": "Dealtype",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/PipelineStage"
            },
            "title": "Stages",
            "type": "array"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "dealType",
          "stages"
        ],
        "title": "PipelineTemplateRes",
        "type": "object"
      },
      "PipelineTemplateStatsRes": {
        "properties": {
          "dealType": {
            "enum": [
              "BRRRR",
              "FLIP"
            ],
            "title": "Dealtype",
            "type": "string"
          },
          "orphanStageDealCount": {
            "default": 0,
            "title": "Orphanstagedealcount",
            "type": "integer"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/PipelineStageStat"
            },
            "title": "Stages",
            "type": "array"
          }
        },
        "required": [
          "dealType",
          "stages"
        ],
        "title": "PipelineTemplateStatsRes",
        "type": "object"
      },
      "PipelineTemplateUpsert": {
        "description": "Payload for PUT /pipeline-templates/{deal_type}.",
        "properties": {
          "stages": {
            "items": {
              "$ref": "#/components/schemas/PipelineStage"
            },
            "minItems": 1,
            "title": "Stages",
            "type": "array"
          }
        },
        "required": [
          "stages"
        ],
        "title": "PipelineTemplateUpsert",
        "type": "object"
      },
...
      "RentComp-Output": {
        "properties": {
          "rent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rent"
          },
          "time_on_market": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time On Market"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "RentComp",
        "type": "object"
      },
...
      "SoldComp-Output": {
        "properties": {
          "arv": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arv"
          },
          "how_long_ago": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "How Long Ago"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "SoldComp",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/bought-deals": {
      "get": {
        "operationId": "get_bought_deals_bought_deals_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/BoughtBrrrDealRes"
                      },
                      {
                        "$ref": "#/components/schemas/BoughtFlipDealRes"
                      }
                    ]
                  },
                  "title": "Response Get Bought Deals Bought Deals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Bought Deals"
      },
      "post": {
        "operationId": "add_bought_deal_bought_deals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "BRRRR": "#/components/schemas/BoughtBrrrDealCreate",
                    "FLIP": "#/components/schemas/BoughtFlipDealCreate"
                  },
                  "propertyName": "deal_type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/BoughtBrrrDealCreate"
                  },
                  {
                    "$ref": "#/components/schemas/BoughtFlipDealCreate"
                  }
                ],
                "title": "Deal"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BoughtBrrrDealRes"
                    },
                    {
                      "$ref": "#/components/schemas/BoughtFlipDealRes"
                    }
                  ],
                  "title": "Response Add Bought Deal Bought Deals Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Bought Deal"
      }
    },
    "/bought-deals/from-active/{deal_id}": {
      "post": {
        "operationId": "move_to_bought_bought_deals_from_active__deal_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "title": "Deal Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deal_type",
            "required": false,
            "schema": {
              "default": "BRRRR",
              "title": "Deal Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BoughtBrrrDealRes"
                    },
                    {
                      "$ref": "#/components/schemas/BoughtFlipDealRes"
                    }
                  ],
                  "title": "Response Move To Bought Bought Deals From Active  Deal Id  Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move To Bought"
      }
    },
    "/bought-deals/{deal_id}": {
      "delete": {
        "operationId": "delete_bought_deal_bought_deals__deal_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "title": "Deal Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deal_type",
            "required": false,
            "schema": {
              "default": "BRRRR",
              "title": "Deal Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Bought Deal"
      },
      "put": {
        "operationId": "update_bought_deal_bought_deals__deal_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/BoughtBrrrDealCreate"
                  },
                  {
                    "$ref": "#/components/schemas/BoughtFlipDealCreate"
                  }
                ],
                "title": "Deal"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BoughtBrrrDealRes"
                    },
                    {
                      "$ref": "#/components/schemas/BoughtFlipDealRes"
                    }
                  ],
                  "title": "Response Update Bought Deal Bought Deals  Deal Id  Put"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Bought Deal"
      }
    },
...
    "/pipeline-templates/{deal_type}": {
      "put": {
        "operationId": "update_pipeline_template_route_pipeline_templates__deal_type__put",
        "parameters": [
          {
            "in": "path",
            "name": "deal_type",
            "required": true,
            "schema": {
              "title": "Deal Type",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineTemplateUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineTemplateRes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Pipeline Template Route"
      }
    },
    "/pipeline-templates/{deal_type}/stats": {
      "get": {
        "operationId": "pipeline_template_stats_route_pipeline_templates__deal_type__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "deal_type",
            "required": true,
            "schema": {
              "title": "Deal Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineTemplateStatsRes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pipeline Template Stats Route"
      }
    },
...
  }
}